refactor: move managers to new folder (#17929)
