+++ web/src/lib/managers/event-manager.svelte.ts
